requirements_enhanced.txt
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.8.0
requests>=2.31.0
pillow>=10.1.0
openai>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
